отрефакторил LoginPage
