--- spotify-react-app/src/pages/LoginPage/components/AccountSignupBox/AccountSignupBox.tsx
@@ -1,12 +1,0 @@
-import "./style.css";
-
-const AccountSignupBox = () => {
-  return (
-    <div className="signup-section">
-      <p className="signup-prompt-text-style">Don't have an account?</p>
-      <button className="sign-up-button-style">Sign up for Spotify</button>
-    </div>
-  );
-}
-
-export default AccountSignupBox;
