--- spotify-react-app/src/pages/LoginPage/components/AccountSignupBox/style.css
@@ -1,37 +1,0 @@
-.signup-section {
-    box-sizing: border-box;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    margin-top: 25px;
-}
-
-.signup-prompt-text-style {
-    flex: 0 0 auto;
-    font-weight: 700;
-    font-size: 17px;
-    font-family: "Inter", sans-serif;
-    align-self: center;
-    color: black;
-    white-space: pre-wrap;
-}
-
-.sign-up-button-style {
-    flex: 0 0 auto;
-    border: 1px solid gray;
-    border-radius: 500px;
-    font-weight: 700;
-    font-size: 13px;
-    font-family: "Inter", sans-serif;
-    background: transparent;
-    box-sizing: border-box;
-    display: block;
-    height: 48px;
-    margin: 10px 0;
-    color: dimgray;
-    text-transform: uppercase;
-    letter-spacing: 1.5px;
-    cursor: pointer;
-    width: 100%;
-}
