--- spotify-react-app/src/pages/LoginPage/components/AuthForm/icons/SvgIcon1.tsx
@@ -1,21 +1,0 @@
-function SvgIcon1({ className }: { className: string }) {
-  return (
-    <div className={className}>
-      <svg fill="none" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg" x="0" y="0">
-        <g id="Frame_6" clipPath="url(#clip1_0_878)" xmlns="http://www.w3.org/2000/svg" data-node-id="0:888">
-          <g id="Frame_7" data-node-id="0:889">
-            <path
-              id="Vector_2"
-              d="M22,12.0187c0,-5.523 -4.477,-10 -10,-10c-5.523,0 -10,4.477 -10,10c0,4.991 3.657,9.128 8.438,9.878v-6.987h-2.54v-2.891h2.54v-2.203c0,-2.506 1.492,-3.89 3.777,-3.89c1.093,0 2.238,0.195 2.238,0.195v2.46h-1.26c-1.243,0 -1.63,0.771 -1.63,1.563v1.875h2.773l-0.443,2.89h-2.33v6.988c4.78,-0.75 8.437,-4.887 8.437,-9.878z"
-              fill="currentColor"
-              fillOpacity="0.9"
-              data-node-id="0:890"
-            />
-          </g>
-        </g>
-      </svg>
-    </div>
-  );
-}
-
-export default SvgIcon1;
