--- spotify-react-app/src/pages/LoginPage/components/AuthForm/icons/SvgIcon2.tsx
@@ -1,22 +1,0 @@
-function SvgIcon2({ className }: { className: string }) {
-  return (
-    <div className={className}>
-      <svg fill="none" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg" x="0" y="0">
-        <g id="Frame_8" xmlns="http://www.w3.org/2000/svg" data-node-id="0:894">
-          <g id="Frame_9" data-node-id="0:895">
-            <g id="Group" data-node-id="0:896">
-              <path
-                id="Vector_3"
-                d="M12.1442,6.8927c0.858,0 1.933,-0.58 2.573,-1.353c0.58,-0.701 1.002,-1.679 1.002,-2.658c0,-0.132 -0.012,-0.265 -0.036,-0.374c-0.954,0.036 -2.102,0.64 -2.79,1.449c-0.544,0.616 -1.039,1.583 -1.039,2.573c0,0.145 0.024,0.29 0.036,0.338c0.061,0.013 0.157,0.025 0.254,0.025zM9.1242,21.5077c1.172,0 1.691,-0.785 3.153,-0.785c1.486,0 1.812,0.761 3.116,0.761c1.281,0 2.138,-1.184 2.947,-2.343c0.906,-1.329 1.281,-2.634 1.305,-2.694c-0.085,-0.024 -2.537,-1.027 -2.537,-3.841c0,-2.44 1.933,-3.539 2.042,-3.624c-1.281,-1.836 -3.225,-1.884 -3.757,-1.884c-1.437,0 -2.609,0.87 -3.346,0.87c-0.797,0 -1.848,-0.822 -3.092,-0.822c-2.367,0 -4.771,1.957 -4.771,5.653c0,2.295 0.894,4.723 1.993,6.293c0.942,1.329 1.764,2.416 2.947,2.416z"
-                fill="currentColor"
-                data-node-id="0:897"
-              />
-            </g>
-          </g>
-        </g>
-      </svg>
-    </div>
-  );
-}
-
-export default SvgIcon2;
