+++ spotify-react-app/src/pages/LoginPage/components/InputWidget/InputWidget.css
@@ -1,3 +1,41 @@
+.login-form-label-container {
+    flex: 0 0 auto;
+    margin-top: 29px;
+}
+
+.input-label-container {
+    border: none;
+    box-sizing: border-box;
+    flex-direction: row;
+    align-items: center;
+    justify-content: start;
+    width: 100%;
+    height: 74px;
+}
+
+.email-label-style {
+    font-weight: 700;
+    font-size: 13px;
+    font-family: "Inter", sans-serif;
+    display: block;
+    margin-bottom: 10px;
+    color: black;
+}
+
+.email-input-style {
+    border: none;
+    border-radius: 4px;
+    font-weight: 400;
+    font-size: 15px;
+    font-family: "Inter", sans-serif;
+    background: white;
+    box-sizing: border-box;
+    width: 100%;
+    height: 48px;
+    padding-left: 13px;
+    box-shadow: inset 0 0 0 1px #878787;
+}
+
 .password-field-container {
     flex: 0 0 auto;
     margin-top: 18px;
@@ -58,3 +96,4 @@
     position: absolute;
     right: 0;
 }
+
